Update includes in moved files
